Make it compile cleanly (except XSValueTest..) with gcc -Wall -Wextra
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@407876 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/framework/XMLFormatter.cpp
@@ -505,8 +505,7 @@ XMLFormatter::handleUnEscapedChars(const XMLCh *                  srcPtr,
    unsigned int count = oCount; 
  
    while (count) { 
-      const unsigned srcChars  
-         = count > kTmpBufSize ? kTmpBufSize : count; 
+      const unsigned srcChars = (count > kTmpBufSize) ? (int)kTmpBufSize : count; 
  
       const unsigned int outBytes  
          = fXCoder->transcodeTo(srcPtr, srcChars,
